refactor(quic): separate packet transport and tests
